Move utils, internal, and context files
