+++ src/components/canvas/context.ts
@@ -1,4 +1,4 @@
-import { createContext } from '../../create-context.js';
+import { createContext } from '../../context/create-context.js';
 import type { CXCanvasContext, CXRenderingContext } from '../../types.js';
 
 export const canvasContext = createContext<CXCanvasContext>();
