+++ src/internal/global.ts
@@ -1,4 +1,4 @@
-import type { CXGlobal } from './types.js';
+import type { CXGlobal } from '../types.js';
 
 export const cxGlobal: CXGlobal = {
   hookIndex: 0,
