+++ src/utils/type-guards.ts
@@ -1,4 +1,4 @@
-import type { AnyArray } from './types.js';
+import type { AnyArray } from '../types.js';
 
 export const isArray = <T>(input: T): input is Extract<T, AnyArray> =>
   Array.isArray(input);
